Fixes a couple of issues to add fp16 training support (#488)
* Fixes a couple of issues to add fp16 training support (#476)

* Add half precision support to `nanodet_plus` head

Moves the explicit `sigmoid` calculation inside the `dsl_assigner` so
that `binary_cross_entropy_with_logits` can be used. This allows for the
use of `auto_cast` to support training with `fp16` precision. If this is
not done `torch` will complain that using `binary_cross_entropy` with
`fp16` is unstable and as such refuses to train the model in `fp16`
precision.

* Add model precision settings to config

Allows for setting the model precision during training using the config
system.

Co-authored-by: RangiLyu <lyuchqi@gmail.com>

* fix lint

* lightning version

Co-authored-by: Bjarne <crisp-snakey@users.noreply.github.com>